refactor: extract theme builder
